+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 5.0.4 (unreleased)
 
 - Added `max_result_window` option
+- Added `bulk_queue_name` option
 
 ## 5.0.3 (2022-03-13)